refactor: member/auth 디렉토리 추가
